backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
gitpython==3.1.40
elasticsearch==8.11.0
httpx==0.25.2
PyGithub==2.1.1
mcp>=1.0.0
pydantic>=2.0.0
